requirements.txt
python-telegram-bot[job-queue]
playwright
